Refactor client knowledge management code
